+++ lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/dao/PartitionInfoDao.java
@@ -6,9 +6,9 @@ package com.dmetasoul.lakesoul.meta.dao;
 
 import com.dmetasoul.lakesoul.meta.DBConnector;
 import com.dmetasoul.lakesoul.meta.DBUtil;
-import com.dmetasoul.lakesoul.meta.entity.CommitOp;
-import com.dmetasoul.lakesoul.meta.entity.PartitionInfo;
-import com.dmetasoul.lakesoul.meta.rbac.AuthZEnforcer;
+import com.dmetasoul.lakesoul.meta.entity.*;
+import com.dmetasoul.lakesoul.meta.jnr.NativeMetadataJavaClient;
+import com.dmetasoul.lakesoul.meta.jnr.NativeUtils;
 
 import java.sql.*;
 import java.util.*;
@@ -17,6 +17,12 @@ import java.util.stream.Collectors;
 public class PartitionInfoDao {
 
     public void insert(PartitionInfo partitionInfo) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            Integer count = NativeMetadataJavaClient.insert(
+                    NativeUtils.CodedDaoType.InsertPartitionInfo,
+                    JniWrapper.newBuilder().addPartitionInfo(partitionInfo).build());
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         try {
@@ -32,6 +38,11 @@ public class PartitionInfoDao {
     }
 
     public boolean transactionInsert(List<PartitionInfo> partitionInfoList, List<String> snapshotList) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            if (partitionInfoList.isEmpty()) return true;
+            Integer count = NativeMetadataJavaClient.insert(NativeUtils.CodedDaoType.TransactionInsertPartitionInfo, JniWrapper.newBuilder().addAllPartitionInfo(partitionInfoList).build());
+            return count > 0;
+        }
         boolean flag = true;
         Connection conn = null;
         PreparedStatement pstmt = null;
@@ -73,7 +84,7 @@ public class PartitionInfoDao {
 
     private void insertSinglePartitionInfo(Connection conn, PreparedStatement pstmt, PartitionInfo partitionInfo)
             throws SQLException {
-        Array array = conn.createArrayOf("UUID", partitionInfo.getSnapshotList().toArray());
+        Array array = conn.createArrayOf("UUID", partitionInfo.getSnapshotList().stream().map(DBUtil::toJavaUUID).toArray());
         pstmt.setString(1, partitionInfo.getTableId());
         pstmt.setString(2, partitionInfo.getPartitionDesc());
         pstmt.setInt(3, partitionInfo.getVersion());
@@ -85,6 +96,12 @@ public class PartitionInfoDao {
     }
 
     public void deleteByTableIdAndPartitionDesc(String tableId, String partitionDesc) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            Integer count = NativeMetadataJavaClient.update(
+                    NativeUtils.CodedDaoType.DeletePartitionInfoByTableIdAndPartitionDesc,
+                    Arrays.asList(tableId, partitionDesc));
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         String sql =
@@ -102,6 +119,10 @@ public class PartitionInfoDao {
     }
 
     public void deleteByTableId(String tableId) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            Integer count = NativeMetadataJavaClient.update(NativeUtils.CodedDaoType.DeletePartitionInfoByTableId, Collections.singletonList(tableId));
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         String sql = "delete from partition_info where table_id = ? ";
@@ -118,6 +139,12 @@ public class PartitionInfoDao {
     }
 
     public void deletePreviousVersionPartition(String tableId, String partitionDesc, long utcMills) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            Integer count = NativeMetadataJavaClient.update(
+                    NativeUtils.CodedDaoType.DeletePreviousVersionPartition,
+                    Arrays.asList(tableId, partitionDesc, Long.toString(utcMills)));
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         String sql = "delete from partition_info where table_id = ? and partition_desc = ? and timestamp <= ?";
@@ -136,6 +163,15 @@ public class PartitionInfoDao {
     }
 
     public List<PartitionInfo> findByTableIdAndParList(String tableId, List<String> partitionDescList) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            if (partitionDescList.isEmpty()) return Collections.emptyList();
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.ListPartitionDescByTableIdAndParList,
+                    Arrays.asList(tableId,
+                            String.join(NativeUtils.PARTITION_DESC_DELIM, partitionDescList)));
+            if (jniWrapper == null) return null;
+            return jniWrapper.getPartitionInfoList();
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -176,6 +212,14 @@ public class PartitionInfoDao {
     }
 
     public PartitionInfo selectLatestPartitionInfo(String tableId, String partitionDesc) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.SelectOnePartitionVersionByTableIdAndDesc,
+                    Arrays.asList(tableId, partitionDesc));
+            if (jniWrapper == null) return null;
+            List<PartitionInfo> partitionInfoList = jniWrapper.getPartitionInfoList();
+            return partitionInfoList.isEmpty() ? null : partitionInfoList.get(0);
+        }
         String sql = String.format(
                 "select m.table_id, t.partition_desc, m.version, m.commit_op, m.snapshot, m.expression, m.domain from (" +
                         "select table_id,partition_desc,max(version) from partition_info " +
@@ -186,7 +230,22 @@ public class PartitionInfoDao {
         return getPartitionInfo(sql);
     }
 
-    public long getLastedTimestamp(String tableId, String partitionDesc) {
+    public long getLatestTimestamp(String tableId, String partitionDesc) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            if (null == partitionDesc || "".equals(partitionDesc)) {
+                List<String> result = NativeMetadataJavaClient.queryScalar(
+                        NativeUtils.CodedDaoType.GetLatestTimestampFromPartitionInfoWithoutPartitionDesc,
+                        Collections.singletonList(tableId));
+                if (result == null || result.isEmpty()) return -1;
+                return Long.parseLong(result.get(0));
+            } else {
+                List<String> result = NativeMetadataJavaClient.queryScalar(
+                        NativeUtils.CodedDaoType.GetLatestTimestampFromPartitionInfo,
+                        Arrays.asList(tableId, partitionDesc));
+                if (result == null || result.isEmpty()) return -1;
+                return Long.parseLong(result.get(0));
+            }
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -217,6 +276,14 @@ public class PartitionInfoDao {
     }
 
     public int getLastedVersionUptoTime(String tableId, String partitionDesc, long utcMills) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            List<String> result = NativeMetadataJavaClient.queryScalar(
+                    NativeUtils.CodedDaoType.GetLatestVersionUpToTimeFromPartitionInfo,
+                    Arrays.asList(tableId, partitionDesc, Long.toString(utcMills)));
+            System.out.println(result);
+            if (result == null || result.isEmpty()) return -1;
+            return Integer.parseInt(result.get(0));
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -247,6 +314,14 @@ public class PartitionInfoDao {
     }
 
     public long getLastedVersionTimestampUptoTime(String tableId, String partitionDesc, long utcMills) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            List<String> result = NativeMetadataJavaClient.queryScalar(
+                    NativeUtils.CodedDaoType.GetLatestVersionTimestampUpToTimeFromPartitionInfo,
+                    Arrays.asList(tableId, partitionDesc, Long.toString(utcMills)));
+            System.out.println(result);
+            if (result == null || result.isEmpty()) return 0;
+            return Long.parseLong(result.get(0));
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -277,6 +352,13 @@ public class PartitionInfoDao {
     }
 
     public List<PartitionInfo> getPartitionVersions(String tableId, String partitionDesc) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.ListPartitionByTableIdAndDesc,
+                    Arrays.asList(tableId, partitionDesc));
+            if (jniWrapper == null) return null;
+            return jniWrapper.getPartitionInfoList();
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -300,6 +382,11 @@ public class PartitionInfoDao {
     }
 
     public List<PartitionInfo> getPartitionDescByTableId(String tableId) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(NativeUtils.CodedDaoType.ListPartitionByTableId, Collections.singletonList(tableId));
+            if (jniWrapper == null) return null;
+            return jniWrapper.getPartitionInfoList();
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -327,7 +414,14 @@ public class PartitionInfoDao {
     }
 
     public PartitionInfo findByKey(String tableId, String partitionDesc, int version) {
-
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.SelectPartitionVersionByTableIdAndDescAndVersion,
+                    Arrays.asList(tableId, partitionDesc, Integer.toString(version)));
+            if (jniWrapper == null) return null;
+            List<PartitionInfo> partitionInfoList = jniWrapper.getPartitionInfoList();
+            return partitionInfoList.isEmpty() ? null : partitionInfoList.get(0);
+        }
         String sql = String.format(
                 "select * from partition_info where table_id = '%s' and partition_desc = '%s' and version = %d",
                 tableId, partitionDesc, version);
@@ -336,6 +430,13 @@ public class PartitionInfoDao {
 
     public List<PartitionInfo> getPartitionsFromVersion(String tableId, String partitionDesc, int startVersion,
                                                         int endVersion) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.ListPartitionVersionByTableIdAndPartitionDescAndVersionRange,
+                    Arrays.asList(tableId, partitionDesc, Integer.toString(startVersion), Integer.toString(endVersion)));
+            if (jniWrapper == null) return null;
+            return jniWrapper.getPartitionInfoList();
+        }
         String sql = String.format(
                 "select * from partition_info where table_id = '%s' and partition_desc = '%s' and version >= %d and " +
                         "version <= %d",
@@ -344,6 +445,13 @@ public class PartitionInfoDao {
     }
 
     public List<PartitionInfo> getOnePartition(String tableId, String partitionDesc) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.SelectOnePartitionVersionByTableIdAndDesc,
+                    Arrays.asList(tableId, partitionDesc));
+            if (jniWrapper == null) return null;
+            return jniWrapper.getPartitionInfoList();
+        }
         String sql =
                 String.format("select * from partition_info where table_id = '%s' and partition_desc = '%s' limit 1",
                         tableId, partitionDesc);
@@ -352,6 +460,13 @@ public class PartitionInfoDao {
 
     public List<PartitionInfo> getPartitionsFromTimestamp(String tableId, String partitionDesc, long startTimestamp,
                                                           long endTimestamp) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.ListPartitionVersionByTableIdAndPartitionDescAndTimestampRange,
+                    Arrays.asList(tableId, partitionDesc, Long.toString(startTimestamp), Long.toString(endTimestamp)));
+            if (jniWrapper == null) return null;
+            return jniWrapper.getPartitionInfoList();
+        }
         String sql = String.format(
                 "select * from partition_info where table_id = '%s' and partition_desc = '%s' and timestamp >= %d and" +
                         " timestamp < %d",
@@ -360,6 +475,13 @@ public class PartitionInfoDao {
     }
 
     public List<String> getAllPartitionDescByTableId(String tableId) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.ListPartitionByTableId,
+                    Collections.singletonList(tableId));
+            if (jniWrapper == null) return null;
+            return jniWrapper.getPartitionInfoList().stream().map(PartitionInfo::getPartitionDesc).collect(Collectors.toList());
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -382,6 +504,13 @@ public class PartitionInfoDao {
 
     public Set<CommitOp> getCommitOpsBetweenVersions(String tableId, String partitionDesc, int firstVersion,
                                                      int secondVersion) {
+        if (NativeUtils.NATIVE_METADATA_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.ListCommitOpsBetweenVersions,
+                    Arrays.asList(tableId, partitionDesc, Integer.toString(firstVersion), Long.toString(secondVersion)));
+            if (jniWrapper == null) return null;
+            return jniWrapper.getPartitionInfoList().stream().map(PartitionInfo::getCommitOp).collect(Collectors.toSet());
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -473,7 +602,7 @@ public class PartitionInfoDao {
                 .setDomain(rs.getString("domain"))
                 .setTimestamp(rs.getLong("timestamp"));
         Array snapshotArray = rs.getArray("snapshot");
-        partitionInfo.addAllSnapshot(Arrays.stream((UUID[]) snapshotArray.getArray()).map(UUID::toString).collect(Collectors.toList()));
+        partitionInfo.addAllSnapshot(Arrays.stream((UUID[]) snapshotArray.getArray()).map(DBUtil::toProtoUuid).collect(Collectors.toList()));
         partitionInfo.setExpression(rs.getString("expression"));
         return partitionInfo.build();
     }
@@ -485,7 +614,7 @@ public class PartitionInfoDao {
                 .setVersion(rs.getInt("version"))
                 .setCommitOp(CommitOp.valueOf(rs.getString("commit_op")));
         Array snapshotArray = rs.getArray("snapshot");
-        partitionInfo.addAllSnapshot(Arrays.stream((UUID[]) snapshotArray.getArray()).map(UUID::toString).collect(Collectors.toList()));
+        partitionInfo.addAllSnapshot(Arrays.stream((UUID[]) snapshotArray.getArray()).map(DBUtil::toProtoUuid).collect(Collectors.toList()));
         partitionInfo.setExpression(rs.getString("expression"));
         return partitionInfo.build();
     }
